+++ packages/xrpl/src/models/methods/accountObjects.ts
@@ -1,5 +1,6 @@
 import {
   AMM,
+  Bridge,
   Check,
   DepositPreauth,
   Escrow,
@@ -8,12 +9,15 @@ import {
   RippleState,
   SignerList,
   Ticket,
+  XChainOwnedClaimID,
+  XChainOwnedCreateAccountClaimID,
 } from '../ledger'
 
 import { BaseRequest, BaseResponse, LookupByLedgerRequest } from './baseMethod'
 
 export type AccountObjectType =
   | 'amm'
+  | 'bridge'
   | 'check'
   | 'deposit_preauth'
   | 'escrow'
@@ -23,6 +27,8 @@ export type AccountObjectType =
   | 'signer_list'
   | 'state'
   | 'ticket'
+  | 'xchain_owned_create_account_claim_id'
+  | 'xchain_owned_claim_id'
 
 /**
  * The account_objects command returns the raw ledger format for all objects
@@ -67,6 +73,7 @@ export interface AccountObjectsRequest
  */
 export type AccountObject =
   | AMM
+  | Bridge
   | Check
   | DepositPreauth
   | Escrow
@@ -75,6 +82,8 @@ export type AccountObject =
   | SignerList
   | RippleState
   | Ticket
+  | XChainOwnedClaimID
+  | XChainOwnedCreateAccountClaimID
 
 /**
  * Response expected from an {@link AccountObjectsRequest}.