+++ packages/xrpl/HISTORY.md
@@ -4,11 +4,14 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ## Unreleased
 
+## 2.12.0 (2023-09-27)
 ### Added
 * Added `ports` field to `ServerInfoResponse`
+* Support for the XChainBridge amendment.
 
 ### Fixed
 * Fix request model fields related to AMM
+* Rename `AMMAccount` to `Account` on `AMM` ledger objects
 * Fixed `EscrowCancel` and `EscrowFinish` validation
 
 ## 2.11.0 (2023-08-24)